+++ lib/constants.dart
@@ -1,0 +1,10 @@
+// BSD 3-Clause License
+// Copyright (c) 2022, GM Consult Pty Ltd
+// All rights reserved.
+// Porter stemmer algorithm is Copyright (c) 2001, Dr Martin Porter, and
+// Copyright (c) 2002, Richard Boulton, all rights reserved.
+
+/// All the constants used by the Porter2Stemmer implementation classes.
+library constants;
+
+export 'src/porter_2_stemmer_constants.dart' show Porter2StemmerConstants;