+++ lib/extensions.dart
@@ -1,0 +1,13 @@
+// BSD 3-Clause License
+// Copyright (c) 2022, GM Consult Pty Ltd
+// All rights reserved.
+// Porter stemmer algorithm is Copyright (c) 2001, Dr Martin Porter, and
+// Copyright (c) 2002, Richard Boulton, all rights reserved.
+
+/// Reduce a word to its root form using the Porter English stemming algorithm
+/// using a String extension.
+///
+/// Also provides stemming extensions.
+library extensions;
+
+export 'src/porter_2_stemmer_extension.dart' show Porter2StemmerExtension;