+++ lib/porter_2_stemmer.dart
@@ -7,6 +7,10 @@
 /// Reduce a word to its root form using the Porter English stemming algorithm.
 library porter_2_stemmer;
 
-export 'src/porter_2_stemmer_extension.dart' show Porter2StemmerExtension;
+export 'src/porter_2_stemmer_extension.dart' show Porter2StemmerExtensions;
 export 'src/porter_2_stemmer.dart'
-    show Porter2Stemmer, Porter2StemmerBase, Porter2StemmerMixin;
+    show
+        Porter2Stemmer,
+        Porter2StemmerBase,
+        Porter2StemmerMixin,
+        Porter2StemmerExtension;