+++ CHANGELOG.md
@@ -4,6 +4,22 @@ Copyright (c) 2022, GM Consult Pty Ltd
 All rights reserved. 
 -->
 
+### 3.0.0
+**BREAKING CHANGES**
+
+#### *Breaking changes*
+* All constants moved from `Porter2Stemmer` and `Porter2StemmerMixin` to `Porter2StemmerConstants`.
+
+#### *New*
+* New mini library `constants`.
+* New mini library `extensions`.
+
+#### *Updated*
+* Dependencies.
+* Tests.
+* Examples.
+* Documentation.  
+
 ### 2.0.0
 **BREAKING CHANGES**